Updates to chapter 9
