+++ ch09/9.4/plum/build.gradle
@@ -1,9 +1,8 @@
 plugins {                                                      // a
+    id "java"
+    id "application"
     id "com.commercehub.gradle.plugin.avro" version "0.8.0"
 }
-
-apply plugin: 'java'
-apply plugin: 'application'
 
 sourceCompatibility = '1.8'
 
@@ -16,10 +15,10 @@ repositories {
 version = '0.1.0'
 
 dependencies {                                                 // b
-  compile 'org.apache.kafka:kafka-clients:0.9.0.0'
-  compile 'org.apache.avro:avro:1.8.0'
-  compile 'net.sargue:mailgun:1.0.0'
-  compile 'org.slf4j:slf4j-api:1.7.5'
+  compile 'org.apache.kafka:kafka-clients:2.0.0'
+  compile 'org.apache.avro:avro:1.8.2'
+  compile 'net.sargue:mailgun:1.9.0'
+  compile 'org.slf4j:slf4j-api:1.7.25'
 }
 
 avro {
