+++ ch09/9.4/plum/src/main/java/plum/Consumer.java
