+++ ch09/9.4/plum/src/main/java/plum/EchoExecutor.java
@@ -9,16 +9,16 @@ import org.apache.avro.io.*;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.specific.SpecificDatumReader;
 
-import plum.avro.Alert;
+import plum.avro.Alert;                                                // a
 
-public class FullExecutor implements IExecutor {
+public class EchoExecutor implements IExecutor {
 
   private final KafkaProducer<String, String> producer;
   private final String eventsTopic;
 
   private static Schema schema;
   static {
-    try {
+    try {                                                              // b
       schema = new Schema.Parser()
         .parse(EchoExecutor.class.getResourceAsStream("/avro/alert.avsc"));
     } catch (IOException ioe) {
@@ -26,10 +26,9 @@ public class FullExecutor implements IExecutor {
     }
   }
 
-  public FullExecutor(String servers, String eventsTopic) {
+  public EchoExecutor(String servers, String eventsTopic) {
 
-    this.producer = new KafkaProducer(
-      IExecutor.createConfig(servers));
+    this.producer = new KafkaProducer(IExecutor.createConfig(servers));
     this.eventsTopic = eventsTopic;
   }
 
@@ -41,10 +40,9 @@ public class FullExecutor implements IExecutor {
     try {
       Decoder decoder = DecoderFactory.get().jsonDecoder(schema, din);
       DatumReader<Alert> reader = new SpecificDatumReader<Alert>(schema);
-      Alert alert = reader.read(null, decoder);
-      Emailer.send(alert);                                            // a
-      IExecutor.write(this.producer, this.eventsTopic,
-        "{ \"event\": \"email_sent\" }");                             // b
+      Alert alert = reader.read(null, decoder);                        // c
+      System.out.println("Alert " + alert.recipient.name + " about " +
+        alert.notification.summary);                                   // d
     } catch (IOException | AvroTypeException e) {
       System.out.println("Error executing command:" + e.getMessage());
     }
