+++ ch09/9.4/plum/src/main/java/plum/FullExecutor.java
@@ -9,16 +9,16 @@ import org.apache.avro.io.*;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.specific.SpecificDatumReader;
 
-import plum.avro.Alert;                                                // a
+import plum.avro.Alert;
 
-public class EchoExecutor implements IExecutor {
+public class FullExecutor implements IExecutor {
 
   private final KafkaProducer<String, String> producer;
   private final String eventsTopic;
 
   private static Schema schema;
   static {
-    try {                                                              // b
+    try {
       schema = new Schema.Parser()
         .parse(EchoExecutor.class.getResourceAsStream("/avro/alert.avsc"));
     } catch (IOException ioe) {
@@ -26,10 +26,9 @@ public class EchoExecutor implements IExecutor {
     }
   }
 
-  public EchoExecutor(String servers, String eventsTopic) {
+  public FullExecutor(String servers, String eventsTopic) {
 
-    this.producer = new KafkaProducer(
-      IExecutor.createConfig(servers));
+    this.producer = new KafkaProducer(IExecutor.createConfig(servers));
     this.eventsTopic = eventsTopic;
   }
 
@@ -41,9 +40,10 @@ public class EchoExecutor implements IExecutor {
     try {
       Decoder decoder = DecoderFactory.get().jsonDecoder(schema, din);
       DatumReader<Alert> reader = new SpecificDatumReader<Alert>(schema);
-      Alert alert = reader.read(null, decoder);                        // c
-      System.out.println("Alert " + alert.recipient.name + " about " +
-        alert.notification.summary);                                   // d
+      Alert alert = reader.read(null, decoder);
+      Emailer.send(alert);                                            // a
+      IExecutor.write(this.producer, this.eventsTopic,
+        "{ \"event\": \"email_sent\" }");                             // b
     } catch (IOException | AvroTypeException e) {
       System.out.println("Error executing command:" + e.getMessage());
     }
