+++ ch09/9.4/plum/src/main/resources/avro/alert.avsc
